Move webgl-specific types into addon
