+++ addons/addon-webgl/src/CharAtlasCache.ts
@@ -6,7 +6,7 @@
 import { TextureAtlas } from './TextureAtlas';
 import { ITerminalOptions, Terminal } from '@xterm/xterm';
 import { ITerminal, ReadonlyColorSet } from 'browser/Types';
-import { ICharAtlasConfig, ITextureAtlas } from 'browser/renderer/shared/Types';
+import { ICharAtlasConfig, ITextureAtlas } from './Types';
 import { generateConfig, configEquals } from 'CharAtlasUtils';
 
 interface ITextureAtlasCacheEntry {
