+++ addons/addon-webgl/src/renderLayer/BaseRenderLayer.ts
@@ -5,7 +5,7 @@
 
 import { ReadonlyColorSet } from 'browser/Types';
 import { acquireTextureAtlas } from 'CharAtlasCache';
-import { IRenderDimensions, ITextureAtlas } from 'browser/renderer/shared/Types';
+import { IRenderDimensions } from 'browser/renderer/shared/Types';
 import { ICoreBrowserService, IThemeService } from 'browser/services/Services';
 import { Disposable, toDisposable } from 'vs/base/common/lifecycle';
 import { CellData } from 'common/buffer/CellData';
@@ -14,6 +14,7 @@ import { Terminal } from '@xterm/xterm';
 import { IRenderLayer } from './Types';
 import { throwIfFalsy } from 'browser/renderer/shared/RendererUtils';
 import { TEXT_BASELINE } from 'Constants';
+import type { ITextureAtlas } from '../Types';
 
 export abstract class BaseRenderLayer extends Disposable implements IRenderLayer {
   private _canvas: HTMLCanvasElement;
