+++ src/browser/renderer/shared/README.md
@@ -1,1 +1,1 @@
-This folder contains files that are shared between the renderer addons, but not necessarily bundled into the `xterm` module.
+This folder contains files that are shared between the renderers.
